# Moving LogRecord implementation to SDK