--- a/opentelemetry-appender-log/src/lib.rs
+++ b/opentelemetry-appender-log/src/lib.rs
@@ -130,7 +130,7 @@ where
             log_record.set_severity_number(severity_of_level(record.level()));
             log_record.set_severity_text(record.level().as_str().into());
             log_record.set_body(AnyValue::from(record.args().to_string()));
-            log_record.set_attributes(log_attributes(record.key_values()));
+            log_record.add_attributes(log_attributes(record.key_values()));
 
             self.logger.emit(log_record);
         }