--- a/opentelemetry-sdk/benches/log.rs
+++ b/opentelemetry-sdk/benches/log.rs
@@ -4,12 +4,14 @@ use std::time::SystemTime;
 use async_trait::async_trait;
 use criterion::{criterion_group, criterion_main, Criterion};
 
-use opentelemetry::logs::{AnyValue, LogRecord, Logger as _, LogResult, LoggerProvider as _, Severity};
+use opentelemetry::logs::{
+    AnyValue, LogRecord, LogRecordBuilder, LogResult, Logger as _, LoggerProvider as _, Severity,
+};
 use opentelemetry::trace::Tracer;
 use opentelemetry::trace::TracerProvider as _;
 use opentelemetry::Key;
 use opentelemetry_sdk::export::logs::{LogData, LogExporter};
-use opentelemetry_sdk::logs::{LoggerProvider, Logger};
+use opentelemetry_sdk::logs::{Logger, LoggerProvider};
 use opentelemetry_sdk::trace::{config, Sampler, TracerProvider};
 
 #[derive(Debug)]
@@ -22,7 +24,7 @@ impl LogExporter for VoidExporter {
     }
 }
 
-fn log_benchmark_group<F: Fn(& Logger)>(c: &mut Criterion, name: &str, f: F) {
+fn log_benchmark_group<F: Fn(&Logger)>(c: &mut Criterion, name: &str, f: F) {
     let mut group = c.benchmark_group(name);
 
     group.bench_function("no-context", |b| {
@@ -59,14 +61,19 @@ fn log_benchmark_group<F: Fn(& Logger)>(c: &mut Criterion, name: &str, f: F) {
 
 fn criterion_benchmark(c: &mut Criterion) {
     log_benchmark_group(c, "simple-log", |logger| {
-        let record :&_ = logger.create_log_record().set_body(Some("simple log".into()));
-        logger.emit(*record);
+        logger.emit(
+            logger
+                .create_log_record()
+                .with_body("simple log".into())
+                .build(),
+        );
     });
 
     log_benchmark_group(c, "simple-log-with-int", |logger| {
         logger.emit(
-            LogRecord::builder()
-                .with_body("simple log")
+            logger
+                .create_log_record()
+                .with_body("simple log".into())
                 .with_attribute("testint", 2)
                 .build(),
         )
@@ -74,12 +81,9 @@ fn criterion_benchmark(c: &mut Criterion) {
 
     log_benchmark_group(c, "simple-log-with-double", |logger| {
         logger.emit(
-            logger.create_log_record()
-                .with_body("simple log")
-                .with_attribute("testdouble", 2.2)
-                .build(),
-            LogRecord::builder()
-                .with_body("simple log")
+            logger
+                .create_log_record()
+                .with_body("simple log".into())
                 .with_attribute("testdouble", 2.2)
                 .build(),
         )
@@ -87,8 +91,9 @@ fn criterion_benchmark(c: &mut Criterion) {
 
     log_benchmark_group(c, "simple-log-with-string", |logger| {
         logger.emit(
-            LogRecord::builder()
-                .with_body("simple log")
+            logger
+                .create_log_record()
+                .with_body("simple log".into())
                 .with_attribute("teststring", "test")
                 .build(),
         )
@@ -96,8 +101,9 @@ fn criterion_benchmark(c: &mut Criterion) {
 
     log_benchmark_group(c, "simple-log-with-bool", |logger| {
         logger.emit(
-            LogRecord::builder()
-                .with_body("simple log")
+            logger
+                .create_log_record()
+                .with_body("simple log".into())
                 .with_attribute("testbool", AnyValue::Boolean(true))
                 .build(),
         )
@@ -106,8 +112,9 @@ fn criterion_benchmark(c: &mut Criterion) {
     let bytes = AnyValue::Bytes(vec![25u8, 30u8, 40u8]);
     log_benchmark_group(c, "simple-log-with-bytes", |logger| {
         logger.emit(
-            LogRecord::builder()
-                .with_body("simple log")
+            logger
+                .create_log_record()
+                .with_body("simple log".into())
                 .with_attribute("testbytes", bytes.clone())
                 .build(),
         )
@@ -123,8 +130,9 @@ fn criterion_benchmark(c: &mut Criterion) {
     ]);
     log_benchmark_group(c, "simple-log-with-a-lot-of-bytes", |logger| {
         logger.emit(
-            LogRecord::builder()
-                .with_body("simple log")
+            logger
+                .create_log_record()
+                .with_body("simple log".into())
                 .with_attribute("testbytes", bytes.clone())
                 .build(),
         )
@@ -133,8 +141,9 @@ fn criterion_benchmark(c: &mut Criterion) {
     let vec_any_values = AnyValue::ListAny(vec![AnyValue::Int(25), "test".into(), true.into()]);
     log_benchmark_group(c, "simple-log-with-vec-any-value", |logger| {
         logger.emit(
-            LogRecord::builder()
-                .with_body("simple log")
+            logger
+                .create_log_record()
+                .with_body("simple log".into())
                 .with_attribute("testvec", vec_any_values.clone())
                 .build(),
         )
@@ -149,8 +158,9 @@ fn criterion_benchmark(c: &mut Criterion) {
     ]);
     log_benchmark_group(c, "simple-log-with-inner-vec-any-value", |logger| {
         logger.emit(
-            LogRecord::builder()
-                .with_body("simple log")
+            logger
+                .create_log_record()
+                .with_body("simple log".into())
                 .with_attribute("testvec", vec_any_values.clone())
                 .build(),
         )
@@ -163,8 +173,9 @@ fn criterion_benchmark(c: &mut Criterion) {
     ]));
     log_benchmark_group(c, "simple-log-with-map-any-value", |logger| {
         logger.emit(
-            LogRecord::builder()
-                .with_body("simple log")
+            logger
+                .create_log_record()
+                .with_body("simple log".into())
                 .with_attribute("testmap", map_any_values.clone())
                 .build(),
         )
@@ -183,38 +194,41 @@ fn criterion_benchmark(c: &mut Criterion) {
     ]));
     log_benchmark_group(c, "simple-log-with-inner-map-any-value", |logger| {
         logger.emit(
-            LogRecord::builder()
-                .with_body("simple log")
+            logger
+                .create_log_record()
+                .with_body("simple log".into())
                 .with_attribute("testmap", map_any_values.clone())
                 .build(),
         )
     });
 
     log_benchmark_group(c, "long-log", |logger| {
-        logger.emit(LogRecord::builder().with_body("Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Gravida in fermentum et sollicitudin ac orci phasellus. Ullamcorper dignissim cras tincidunt lobortis feugiat vivamus at augue. Magna etiam tempor orci eu. Sed tempus urna et pharetra pharetra massa.").build())
+        logger.emit(logger.create_log_record().with_body("Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Gravida in fermentum et sollicitudin ac orci phasellus. Ullamcorper dignissim cras tincidunt lobortis feugiat vivamus at augue. Magna etiam tempor orci eu. Sed tempus urna et pharetra pharetra massa.".into()).build())
     });
 
     let now = SystemTime::now();
     log_benchmark_group(c, "full-log", |logger| {
         logger.emit(
-            LogRecord::builder()
-                .with_body("full log")
+            logger
+                .create_log_record()
+                .with_body("full log".into())
                 .with_timestamp(now)
                 .with_observed_timestamp(now)
                 .with_severity_number(Severity::Warn)
-                .with_severity_text(Severity::Warn.name())
+                .with_severity_text(Severity::Warn.name().into())
                 .build(),
         )
     });
 
     log_benchmark_group(c, "full-log-with-4-attributes", |logger| {
         logger.emit(
-            LogRecord::builder()
-                .with_body("full log")
+            logger
+                .create_log_record()
+                .with_body("full log".into())
                 .with_timestamp(now)
                 .with_observed_timestamp(now)
                 .with_severity_number(Severity::Warn)
-                .with_severity_text(Severity::Warn.name())
+                .with_severity_text(Severity::Warn.name().into())
                 .with_attribute("name", "my-event-name")
                 .with_attribute("event.id", 20)
                 .with_attribute("user.name", "otel")
@@ -225,12 +239,13 @@ fn criterion_benchmark(c: &mut Criterion) {
 
     log_benchmark_group(c, "full-log-with-9-attributes", |logger| {
         logger.emit(
-            LogRecord::builder()
-                .with_body("full log")
+            logger
+                .create_log_record()
+                .with_body("full log".into())
                 .with_timestamp(now)
                 .with_observed_timestamp(now)
                 .with_severity_number(Severity::Warn)
-                .with_severity_text(Severity::Warn.name())
+                .with_severity_text(Severity::Warn.name().into())
                 .with_attribute("name", "my-event-name")
                 .with_attribute("event.id", 20)
                 .with_attribute("user.name", "otel")
@@ -266,12 +281,13 @@ fn criterion_benchmark(c: &mut Criterion) {
     ];
     log_benchmark_group(c, "full-log-with-attributes", |logger| {
         logger.emit(
-            LogRecord::builder()
-                .with_body("full log")
+            logger
+                .create_log_record()
+                .with_body("full log".into())
                 .with_timestamp(now)
                 .with_observed_timestamp(now)
                 .with_severity_number(Severity::Warn)
-                .with_severity_text(Severity::Warn.name())
+                .with_severity_text(Severity::Warn.name().into())
                 .with_attributes(attributes.clone())
                 .build(),
         )