Fixes #1189 

## Changes

Opening as draft to get the initial feedback. The changes are:
 - Created `LogRecord`  traits in the API.
 - Added corresponding struct `LogRecord` with implementation in the SDK.

**Earlier:**
```rust
        logger.emit(
            LogRecord::builder()
                .with_body("full log")
                .with_timestamp(now)
                .with_observed_timestamp(now)
                .with_severity_number(Severity::Warn)
                .with_severity_text(Severity::Warn.name())
                .with_attribute("name", "my-event-name")
                .with_attribute("event.id", 20)
                .with_attribute("user.name", "otel")
                .with_attribute("user.email", "otel@opentelemetry.io")
                .with_attribute("code.filename", "log.rs")
                .with_attribute("code.filepath", "opentelemetry_sdk/benches/log.rs")
                .with_attribute("code.lineno", 96)
                .with_attribute("code.namespace", "opentelemetry_sdk::benches::log")
                .with_attribute("log.target", "opentelemetry_sdk::benches::log")
                .build(),
        )

```

**With PR:**
```rust
        let mut log_record = logger.create_log_record();
        log_record.set_body("full log".into());
        log_record.set_timestamp(now);
        log_record.set_observed_timestamp(now);
        log_record.set_severity_number(Severity::Warn);
        log_record.set_severity_text(Severity::Warn.name().into());
        log_record.add_attribute("name", "my-event-name");
        log_record.add_attribute("event.id", 20);
        log_record.add_attribute("user.name", "otel");
        log_record.add_attribute("user.email", "otel@opentelemetry.io");
        log_record.add_attribute("code.filename", "log.rs");
        log_record.add_attribute("code.filepath", "opentelemetry_sdk/benches/log.rs");
        log_record.add_attribute("code.lineno", 96);
        log_record.add_attribute("code.namespace", "opentelemetry_sdk::benches::log");
        log_record.add_attribute("log.target", "opentelemetry_sdk::benches::log");
        logger.emit(log_record);
```

Note - with the new design, the log record creation through chaining required cloning eventually while emitting, so haven't implemented the chaining. The tracing appender also doesn't require chaining as all these log data are populated separately.

This allows us to change the internal SDK implementation for LogRecord without breaking the Bridge API. While there are no current perf benefits with the approach, it enables us to add more optimizations in the SDK to optionally allow the serialization of the log records in the export format without any intermediate allocations within the SDK. I plan to add those optimizations in subsequent PRs once this is finalized. 

**Benchmark**

| Test Name                                        | Main                            | PR                             |
|--------------------------------------------------|---------------------------------|--------------------------------|
| simple-log/no-context                           | [105.28 ns]                     | [89.896 ns]                    |
| simple-log/with-context time                    | [115.49 ns]                     | [90.293 ns]                    |
| simple-log-with-int/no-context                  | [141.40 ns]                     | [123.36 ns]                    |
| simple-log-with-int/with-context                | [147.15 ns]                     | [123.61 ns]                    |
| simple-log-with-double/no-context               | [145.43 ns]                     | [123.32 ns]                    |
| simple-log-with-double/with-context             | [153.33 ns]                     | [122.28 ns]                    |
| simple-log-with-string/no-context               | [144.01 ns]                     | [122.61 ns]                    |
| simple-log-with-string/with-context             | [149.85 ns]                     | [123.61 ns]                    |
| simple-log-with-bool/no-context                 | [146.57 ns]                     | [121.77 ns]                    |
| simple-log-with-bool/with-context               | [152.56 ns]                     | [123.12 ns]                    |
| simple-log-with-bytes/no-context                | [161.11 ns]                     | [139.43 ns]                    |
| simple-log-with-bytes/with-context              | [171.69 ns]                     | [140.73 ns]                    |
| simple-log-with-a-lot-of-bytes/no-context       | [164.77 ns]                     | [140.72 ns]                    |
| simple-log-with-a-lot-of-bytes/with-context     | [172.50 ns]                     | [142.45 ns]                    |
| simple-log-with-vec-any-value/no-context        | [205.88 ns]                     | [194.93 ns]                    |
| simple-log-with-vec-any-value/with-context      | [214.82 ns]                     | [194.06 ns]                    |
| simple-log-with-inner-vec-any-value/no-context  | [275.55 ns]                     | [256.61 ns]                    |
| simple-log-with-inner-vec-any-value/with-context| [280.79 ns]                     | [255.82 ns]                    |
| simple-log-with-map-any-value/no-context        | [234.09 ns]                     | [207.90 ns]                    |
| simple-log-with-map-any-value/with-context      | [244.29 ns]                     | [207.57 ns]                    |
| simple-log-with-inner-map-any-value/no-context  | [335.55 ns]                     | [303.23 ns]                    |
| simple-log-with-inner-map-any-value/with-context| [344.18 ns]                     | [304.65 ns]                    |
| long-log/no-context                             | [106.17 ns]                     | [89.779 ns]                    |
| long-log/with-context                           | [118.93 ns]                     | [91.673 ns]                    |
| full-log/no-context                             | [143.31 ns]                     | [95.529 ns]                    |
| full-log/with-context                           | [145.20 ns]                     | [96.720 ns]                    |
| full-log-with-4-attributes/no-context           | [315.31 ns]                     | [227.49 ns]                    |
| full-log-with-4-attributes/with-context         | [312.81 ns]                     | [235.11 ns]                    |
| full-log-with-9-attributes/no-context           | [548.88 ns]                     | [406.99 ns]                    |
| full-log-with-9-attributes/with-context         | [501.88 ns]                     | [418.26 ns]                    |
| full-log-with-attributes/no-context             | [322.71 ns]                     | [277.63 ns]                    |
| full-log-with-attributes/with-context           | [328.19 ns]                     | [279.08 ns]                    |


**Stress Test** (20 threads)
Main: ~35 million iterations/sec
PR:    ~39 million iterations/sec


## Merge requirement checklist

* [x] [CONTRIBUTING](https://github.com/open-telemetry/opentelemetry-rust/blob/main/CONTRIBUTING.md) guidelines followed
* [x] Unit tests added/updated (if applicable)
* [x] Appropriate `CHANGELOG.md` files updated for non-trivial, user-facing changes
* [x] Changes in public API reviewed (if applicable)