@cijothomas 2024-05-03
> While there are no current perf benefits with the approach,

Can you post benchmarks/stress tests with the changes, to confirm that we are not regressing either. (Just curiosity! I agree that this PR is the right direction)