@lalitb 2024-05-07
> LGTM. Waiting to add changelog, and remove the API to set span/trace context before approving.

changelog is added now. Thanks for the review.